Add files via upload
This was what we started in class on 4/12/2019 with adding the GameObjectClass, and the Handler class. The GameObject class is an abstract class that is the framework for every object. The Handler class creates a LinkedList(A list of the objects) and renders each object. I also add a test object, which is the Box Class